# vpd111.api.com/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Усіх із Різдвом та Новим роком!",
        "description": "Demo my Project ",
        "contact": {
            "email": "admin@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "operationId": "a718f172ff3ac464d723835815f8fb57",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "lastName",
                                    "name",
                                    "phone",
                                    "image",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "image": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "lastName": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Add Category."
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "operationId": "login",
                "requestBody": {
                    "description": "User login data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/oauth/{provider}": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Redirect the user to the OAuth Provider.",
                "operationId": "redirectToProvider",
                "parameters": [
                    {
                        "name": "provider",
                        "in": "path",
                        "description": "OAuth provider name.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to OAuth Provider."
                    }
                }
            }
        },
        "/api/oauth/{provider}/callback": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Obtain the user information from the OAuth Provider.",
                "operationId": "handleProviderCallback",
                "parameters": [
                    {
                        "name": "provider",
                        "in": "path",
                        "description": "OAuth provider name.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User information obtained from the OAuth Provider.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unable to authenticate."
                    }
                }
            }
        },
        "/api/categories": {
            "get": {
                "tags": [
                    "Category"
                ],
                "operationId": "359d1c07dfb50588f7b8fbe2a366fec8",
                "responses": {
                    "200": {
                        "description": "List Categories."
                    }
                }
            }
        },
        "/api/categories/create": {
            "post": {
                "tags": [
                    "Category"
                ],
                "operationId": "c343228f5e1294295e03f7b400761bc9",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "type": "file"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Add Category."
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "operationId": "e41847c3dd46abcd57dd20c4d21b770c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ідентифікатор категорії",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List Categories."
                    },
                    "404": {
                        "description": "Wrong id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong Category Id has been sent. Pls try another one."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "operationId": "e34f793896927e04626d92979d567b66",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ідентифікатор категорії",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успішне видалення категорії"
                    },
                    "404": {
                        "description": "Категорії не знайдено"
                    },
                    "401": {
                        "description": "Не авторизований"
                    }
                }
            }
        },
        "/api/categories/edit/{id}": {
            "post": {
                "tags": [
                    "Category"
                ],
                "operationId": "fcc4fbaaf1008e9b09d850acfc703ece",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ідентифікатор категорії",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "image": {
                                        "type": "file"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Add Category."
                    }
                }
            }
        }
    }
}